bump deps, remove cgo buildtag
no functional changes
